backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
dj-database-url==3.0.1
Django==5.1.5
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
idna==3.10
mysqlclient==2.2.7
numpy==2.1.1
packaging==25.0
pandas==2.2.2
pillow==10.4.0
psycopg==3.2.9
psycopg-binary==3.2.9
pygame==2.6.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2024.1
urllib3==2.3.0
